Move blocks/MovieContent to components
