+++ src/components/MovieContent/Mobile/MobileLayout.js
