+++ src/components/MovieContent/Socials.js
